Separate out kit setup
